+++ jest.setup.js
@@ -1,5 +1,5 @@
-var enzyme = require('enzyme')
-var Adapter = require('enzyme-adapter-react-16')
+const enzyme = require('enzyme')
+const Adapter = require('enzyme-adapter-react-16')
 global.fetch = require('jest-fetch-mock')
 
 enzyme.configure({ adapter: new Adapter() })